zap/format/columns: Format lines in columns.
Zsh’s “print” builtin can print in columns using the “-C” switch.
However, it only uses a simple column-width algorithm that
makes each column as wide as the widest column.

Zap’s “zap/format/column” function, however, uses a variable-column-width
algorithm when printing in columns.  The basic algorithm is the same as
that used in the GNU version of the “ls” command, when issued with the
“-C” switch.

Another noteworthy enhancement is that “zap/format/column” can deal with
terminal codes when calculating the width of the items to print.  This is
done with the same escape sequences used for printing prompts, namely
“%{…%}”.